chore: Refactor environmentItem to separate file
