+++ vercel/resource_project.go
@@ -688,7 +688,7 @@ var nullProject = Project{
 	InstallCommand:  types.StringNull(),
 	OutputDirectory: types.StringNull(),
 	PublicSource:    types.BoolNull(),
-	Environment:     types.SetNull(envVariableElemType),
+	Environment:     types.SetNull(EnvVariableElemType),
 }
 
 func (p *Project) environment(ctx context.Context) ([]EnvironmentItem, error) {
@@ -846,71 +846,6 @@ func (p *Project) toUpdateProjectRequest(ctx context.Context, oldName string) (r
 		GitComments:                          gc.toUpdateProjectRequest(),
 		ResourceConfig:                       resourceConfig.toClientResourceConfig(p.OnDemandConcurrentBuilds, p.BuildMachineType),
 		NodeVersion:                          p.NodeVersion.ValueString(),
-	}, nil
-}
-
-// EnvironmentItem reflects the state terraform stores internally for a project's environment variable.
-type EnvironmentItem struct {
-	Target               types.Set    `tfsdk:"target"`
-	CustomEnvironmentIDs types.Set    `tfsdk:"custom_environment_ids"`
-	GitBranch            types.String `tfsdk:"git_branch"`
-	Key                  types.String `tfsdk:"key"`
-	Value                types.String `tfsdk:"value"`
-	ID                   types.String `tfsdk:"id"`
-	Sensitive            types.Bool   `tfsdk:"sensitive"`
-	Comment              types.String `tfsdk:"comment"`
-}
-
-func (e *EnvironmentItem) equal(other *EnvironmentItem) bool {
-	return e.Key.ValueString() == other.Key.ValueString() &&
-		e.Value.ValueString() == other.Value.ValueString() &&
-		e.Target.Equal(other.Target) &&
-		e.CustomEnvironmentIDs.Equal(other.CustomEnvironmentIDs) &&
-		e.GitBranch.ValueString() == other.GitBranch.ValueString() &&
-		e.Sensitive.ValueBool() == other.Sensitive.ValueBool() &&
-		e.Comment.ValueString() == other.Comment.ValueString()
-}
-
-func (e *EnvironmentItem) toAttrValue() attr.Value {
-	return types.ObjectValueMust(envVariableElemType.AttrTypes, map[string]attr.Value{
-		"id":                     e.ID,
-		"key":                    e.Key,
-		"value":                  e.Value,
-		"target":                 e.Target,
-		"custom_environment_ids": e.CustomEnvironmentIDs,
-		"git_branch":             e.GitBranch,
-		"sensitive":              e.Sensitive,
-		"comment":                e.Comment,
-	})
-}
-
-func (e *EnvironmentItem) toEnvironmentVariableRequest(ctx context.Context) (req client.EnvironmentVariableRequest, diags diag.Diagnostics) {
-	var target []string
-	diags = e.Target.ElementsAs(ctx, &target, true)
-	if diags.HasError() {
-		return req, diags
-	}
-	var customEnvironmentIDs []string
-	diags = e.CustomEnvironmentIDs.ElementsAs(ctx, &customEnvironmentIDs, true)
-	if diags.HasError() {
-		return req, diags
-	}
-
-	var envVariableType string
-	if e.Sensitive.ValueBool() {
-		envVariableType = "sensitive"
-	} else {
-		envVariableType = "encrypted"
-	}
-
-	return client.EnvironmentVariableRequest{
-		Key:                  e.Key.ValueString(),
-		Value:                e.Value.ValueString(),
-		Target:               target,
-		CustomEnvironmentIDs: customEnvironmentIDs,
-		GitBranch:            e.GitBranch.ValueStringPointer(),
-		Type:                 envVariableType,
-		Comment:              e.Comment.ValueString(),
 	}, nil
 }
 
@@ -1185,23 +1120,6 @@ func uncoerceBool(plan, res types.Bool) types.Bool {
 	return res
 }
 
-var envVariableElemType = types.ObjectType{
-	AttrTypes: map[string]attr.Type{
-		"key":   types.StringType,
-		"value": types.StringType,
-		"target": types.SetType{
-			ElemType: types.StringType,
-		},
-		"custom_environment_ids": types.SetType{
-			ElemType: types.StringType,
-		},
-		"git_branch": types.StringType,
-		"id":         types.StringType,
-		"sensitive":  types.BoolType,
-		"comment":    types.StringType,
-	},
-}
-
 var gitCommentsAttrTypes = map[string]attr.Type{
 	"on_commit":       types.BoolType,
 	"on_pull_request": types.BoolType,
@@ -1404,7 +1322,7 @@ func convertResponseToProject(ctx context.Context, response client.ProjectRespon
 			}
 		}
 
-		env = append(env, types.ObjectValueMust(envVariableElemType.AttrTypes, map[string]attr.Value{
+		env = append(env, types.ObjectValueMust(EnvVariableElemType.AttrTypes, map[string]attr.Value{
 			"key":                    types.StringValue(e.Key),
 			"value":                  value,
 			"target":                 targetValue,
@@ -1432,9 +1350,9 @@ func convertResponseToProject(ctx context.Context, response client.ProjectRespon
 		protectionBypassSecret = types.StringValue(plan.ProtectionBypassForAutomationSecret.ValueString())
 	}
 
-	environmentEntry := types.SetValueMust(envVariableElemType, env)
+	environmentEntry := types.SetValueMust(EnvVariableElemType, env)
 	if plan.Environment.IsNull() {
-		environmentEntry = types.SetNull(envVariableElemType)
+		environmentEntry = types.SetNull(EnvVariableElemType)
 	}
 
 	gitComments := types.ObjectNull(gitCommentsAttrTypes)
